Add simulated scan loader that loads from disk for testing
Streaming scans from the serial ports turns out to be very slow.
This is probably due to the fact that there is little time to
plot the scans if the single thread of control (due to GIL) is
busy servicing the serial IO. To verify this, try loading the
scans from disk which should have much higher througput, and should
make the application less IO bound. Results: much faster plotting...